% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_gradient.R
\name{calcul_Gradient}
\alias{calcul_Gradient}
\title{Calcul de Gradient
Une fonction pour calculer une regression logistique avec l'algorithme de la descente de gradient stochastique}
\usage{
calcul_Gradient(X, Type, nb_rows)
}
\arguments{
\item{X, Type, nb_rows}{oui oui}
}
\value{
la fonction renvoie divers éléments
}
\description{
Calcul de Gradient
Une fonction pour calculer une regression logistique avec l'algorithme de la descente de gradient stochastique
}
