% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Logistic Regression}
\usage{
predict(data_test, Train_res, sortie = "Class")
}
\arguments{
\item{data_test}{A test data frame.}

\item{Train_res}{Fitted model from data train using logistic regression.}

\item{sortie}{"Proba" : return probabilities or "Class" : return classes predicted on test data.}
}
\value{
predict function returns a Vector of probabilities or classes predicted on test data.
}
\description{
A function to predict classes using the logistic regression with the stochastic gradient descent algorithm.
}
