% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.DyrRegLog}
\alias{print.DyrRegLog}
\title{Print Override}
\usage{
\method{print}{DyrRegLog}(object)
}
\arguments{
\item{object}{Un objet de classe S3}
}
\value{
la fonction renvoie divers elements
}
\description{
A function to override the print method
}
