% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{Reg_logistic_With_SimpleGrad_TEST}
\alias{Reg_logistic_With_SimpleGrad_TEST}
\title{Logistic Regression}
\usage{
Reg_logistic_With_SimpleGrad_TEST(data_test, Train_res, sortie = "Class")
}
\arguments{
\item{data_test}{A test data frame}

\item{Train_res}{A demander}

\item{sortie}{A demander}
}
\value{
la fonction renvoie divers elements
}
\description{
A function to predict a model using the logistic regression with the stochastic gradient descent algorithm
}
