% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.DyrRegLog}
\alias{summary.DyrRegLog}
\title{Summary Override}
\usage{
\method{summary}{DyrRegLog}(object)
}
\arguments{
\item{object}{Un objet de classe S3}
}
\description{
A function to override the summary method
}
