% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Logistic Regression}
\usage{
fit(
  formula,
  data,
  eta = 0.3,
  iter_Max = 200,
  mode = "Online",
  batch_size = 10,
  tol = 0.001,
  coefs = rep(0, (dim(model.frame(formula, data))[2]) - 1),
  intercept = 0,
  nb_Coeurs = 1
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{eta}{A demander.}

\item{iter_Max}{Maximum iteration realized by the function.}

\item{mode}{The mode to be used in fitting the model (Batch_Simple, Mini_Batch, Online). The default mode is Online.}

\item{batch_size}{Size of the batch.}

\item{tol}{A demander.}

\item{coefs}{A demander.}

\item{intercept}{Logical. Should an intercept be included in the null model ?}

\item{nb_Coeurs}{Number of chores that have to be used by .}
}
\value{
la fonction renvoie divers elements
}
\description{
A function to compute logistic regression with the stochastic gradient descent algorithm
}
