% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Regression logistique
Une fonction pour calculer une regression logistique avec l'algorithme de la descente de gradient stochastique}
\usage{
fit(
  formula,
  data,
  eta = 0.3,
  iter_Max = 200,
  mode = "Online",
  batch_size = 10,
  tol = 0.001,
  coefs = rep(0, (dim(model.frame(formula, data))[2]) - 1),
  intercept = 0,
  nb_Coeurs = 1
)
}
\value{
la fonction renvoie divers elements
}
\description{
Regression logistique
Une fonction pour calculer une regression logistique avec l'algorithme de la descente de gradient stochastique
}
