% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Logistic Regression}
\usage{
predict(data_test, Train_res, sortie = "Class")
}
\arguments{
\item{data_test}{A test data frame.}

\item{Train_res}{Fitted model from data train using logistic regression.}

\item{sortie}{"Proba" : return probabilities or "Class" : return classes predicted on test data.}
}
\value{
Returns a Vector of probabilities or classes predicted on test data.
}
\description{
A function to predict classes using the logistic regression with the stochastic gradient descent algorithm.
}
\examples{
data = iris[(iris$Species=="setosa" | iris$Species=="versicolor"),]
levels(data$Species)
levels(data$Species) <- c(levels(data$Species), c(0,1))
data$Species[data$Species=="setosa"] = 1
data$Species[data$Species=="versicolor"] = 0
data$Species = as.numeric(data$Species)
rows <- sample(nrow(data))
data <- data[rows,]
data$Species = data$Species -4
data_train = data[1:70,]
data_test = data[71:100,]
rm(data)
reg_log=fit(formula=Species~.,data=data_train,eta=0.3,iter_Max=300,mode="Batch_Simple",batch_size=15)
predict(data_test[,-5], reg_log,sortie="Class")
}
